Parcellation Stats to Table

This function calls aparcstats2table to convert parcellation statistics to a table

aparcstats2table(
  subjects,
  outfile = NULL,
  hemi = c("lh", "rh"),
  measure = c("area", "volume", "thickness", "thicknessstd", "meancurv", "gauscurv",
    "foldind", "curvind"),
  sep = c("tab", "space", "comma", "semicolon"),
  parc = c("aparc", "aparc.a2009s"),
  skip = FALSE,
  subj_dir = NULL,
  opts = "",
  verbose = TRUE
)

Arguments

subjects
(character) vector of subjects

outfile
(character) output filename

hemi
(character) hemisphere to run statistics

measure
(character) measure to be calculated

sep
(character) separator for the output file. This will be an attribute of outfile

parc
(character) parcellation to compute on

skip
(logical) if subject does not have parcellation, should the command skip that subject (TRUE) or error (FALSE)

subj_dir
(character path) if a different subjects directory is to be used other than SUBJECTS_DIR from shell, it can be specified here. Use with care as if the command fail, it may not reset the SUBJECTS_DIR back correctly after the error

opts
(character) additional options to aparcstats2table

verbose
(logical) print diagnostic messages

Value

Character filename of output file, with the attribute of the separator

Examples

if (have_fs()) {
   fs_subj_dir()
   outfile = aparcstats2table(subjects = "bert",
                    hemi = "lh",
                    meas = "thickness")
}

#> source '/Applications/freesurfer/FreeSurferEnv.sh' || true ; aparcstats2table --hemi lh --subjects bert --delimiter tab --parc aparc --debug --measure thickness --tablefile /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpoaoS2X/file78e636c630bc.txt 

#> Error in check_fs_result(res = res, fe_before = fe_before, fe_after = fe_after): Command Failed, no output produced!