Read Freesurfer Table Output

This function reads output from a Freesurfer table command, e.g. aparcstats2table, asegstats2table

read_fs_table(file, sep = NULL, stringsAsFactors = FALSE, header = TRUE, ...)

Arguments

file
(character path) filename of text file

sep
separator to override attribute of file, to pass to read.table.

stringsAsFactors
(logical) passed to read.table

header
Is there a header in the data

...
additional arguments to read.table

Value

data.frame from the file

Examples

if (have_fs()) {
   outfile = aparcstats2table(subjects = "bert",
                    hemi = "lh",
                    meas = "thickness")
   df = read_fs_table(outfile)
   seg_outfile = asegstats2table(subjects = "bert", meas = "mean")
   df_seg = read_fs_table(seg_outfile)
}

#> source '/Applications/freesurfer/FreeSurferEnv.sh' || true ; aparcstats2table --hemi lh --subjects bert --delimiter tab --parc aparc --debug --measure thickness --tablefile /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpoaoS2X/file78e6aae495c.txt 

#> Error in check_fs_result(res = res, fe_before = fe_before, fe_after = fe_after): Command Failed, no output produced!

if (FALSE) {
### using the pipe
if (requireNamespace("magrittr", quietly = TRUE)) {
df_seg = asegstats2table(subjects = "bert", meas = "mean") %>% 
            read_fs_table
            }
}