FS Command Wrapper

This function calls Freesurfer command passed to func

fs_cmd(
  func,
  file,
  outfile = NULL,
  retimg = TRUE,
  reorient = FALSE,
  intern = FALSE,
  opts = "",
  verbose = TRUE,
  samefile = FALSE,
  opts_after_outfile = FALSE,
  frontopts = "",
  add_ext = TRUE,
  bin_app = "bin",
  ...
)

Arguments

func

(character) Freesurfer function

file

(character) image to be manipulated

outfile

(character) resultant image name (optional)

retimg

(logical) return image of class nifti

reorient

(logical) If retimg, should file be reoriented when read in? Passed to readnii.

intern

(logical) to be passed to system

opts

(character) operations to be passed to func

verbose

(logical) print out command before running

samefile

(logical) is the output the same file?

opts_after_outfile

(logical) should opts come after the outfile in the Freesurfer command?

frontopts

(character) options/character to put in before filename

add_ext

(logical) should the extension be added to the outfile

bin_app

(character) appendix to add to get_fs

...

additional arguments passed to system.

Value

If retimg then object of class nifti. Otherwise, Result from system command, depends if intern is TRUE or FALSE.