Use Freesurfers mri_surf2surf function to resamples one cortical surface onto another

This function calls Freesurfer mri_surf2surf to resample one cortical surface onto another

mri_surf2surf(
  subject = NULL,
  target_subject = NULL,
  trg_type = c("curv", "w", "mgh", "nii"),
  src_type = c("curv", "w"),
  outfile = NULL,
  hemi = c("lh", "rh"),
  sval = c("thickness"),
  subj_dir = NULL,
  opts = "",
  verbose = TRUE
)

Arguments

subject
(character) vector of subject name

target_subject
(character) vector of target subject name

trg_type
(character) target file type, can be curv, paint (w), mgh, or nii

src_type
(character) source file type, can be curv or paint (w)

outfile
(character) output filename

hemi
(character) hemisphere to run statistics

sval
(character) source file

subj_dir
(character path) if a different subjects directory is to be used other than SUBJECTS_DIR from shell, it can be specified here. Use with care as if the command fail, it may not reset the SUBJECTS_DIR back correctly after the error

opts
(character) additional options to mri_surf2surf

verbose
(logical) print diagnostic messages

Value

Name of output file

Examples

if (have_fs()) {
   out = mri_surf2surf(
   subject = 'bert',
   target_subject = 'fsaverage',
   trg_type  = 'curv',
   src_type  = 'curv',
   hemi = "rh",
   sval = "thickness")
}

#> source '/Applications/freesurfer/FreeSurferEnv.sh' || true ; mri_surf2surf --hemi rh --s bert --trgsubject fsaverage --trg_type curv --sval thickness --src_type curv --tval /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//Rtmp7xlMkZ/file49add7f8cea --debug